[Completes #64477482 Some images in Feed are failing to load]. iOS 6 UI fixes. Ad Hoc profile added.

+++ Morsel/Morsel/Source/Views/MorselScrollView.m
@@ -41,7 +41,7 @@
 
 - (void)scrollToMorsel:(MRSLMorsel *)morsel
 {
-    int morselIndex = [_post.morsels indexOfObject:morsel];
+    NSUInteger morselIndex = [_post.morsels indexOfObject:morsel];
     
     [self scrollRectToVisible:CGRectMake(self.frame.size.width * morselIndex, 0.f, self.frame.size.width, 1.f)
                      animated:NO];
